refactor(table-of-contents): create folder directory of internal components
