+++ src/components/TableOfContents/tableOfContents.utils.ts
@@ -1,0 +1,58 @@
+import { ListProps } from "@chakra-ui/react"
+
+export const customIdRegEx = /^.+(\s*\{#([A-Za-z0-9\-_]+?)\}\s*)$/
+export const emojiRegEx = /<Emoji [^/]+\/>/g
+
+export const slugify = (s: string): string =>
+  encodeURIComponent(String(s).trim().toLowerCase().replace(/\s+/g, "-"))
+
+export const getCustomId = (title: string): string => {
+  const match = customIdRegEx.exec(title)
+  if (match) {
+    return match[2].toLowerCase()
+  }
+  console.warn("Missing custom ID on header: ", title)
+  return slugify(title)
+}
+
+export const trimmedTitle = (title: string): string => {
+  const match = customIdRegEx.exec(title)
+  const trimmedTitle = match ? title.replace(match[1], "").trim() : title
+
+  // Removes Twemoji components from title
+  const emojiMatch = emojiRegEx.exec(trimmedTitle)
+  return emojiMatch ? trimmedTitle.replaceAll(emojiRegEx, "") : trimmedTitle
+}
+
+export interface Item {
+  title: string
+  items?: Array<Item>
+}
+
+/**
+ * Common props used used for the outermost list element in the mobile and desktop renders
+ */
+export const outerListProps: ListProps = {
+  borderStart: "1px solid",
+  borderStartColor: "dropdownBorder",
+  borderTop: 0,
+  fontSize: "sm",
+  lineHeight: 1.6,
+  fontWeight: 400,
+  m: 0,
+  mt: 20,
+  mb: 12,
+  ps: 4,
+  pe: 1,
+  pt: 0,
+  sx: {
+    // TODO: Flip to object syntax with `lg` token after completion of Chakra migration
+    "@media (max-width: 1024px)": {
+      borderStart: 0,
+      borderTop: "1px",
+      borderTopColor: "primary300",
+      ps: 0,
+      pt: 4,
+    },
+  },
+}
